feat(boost): remove join with ping button
The changes remove the "Join w/Ping" button from the boost handlers. This button was not being used and was causing confusion for users, so it has been removed to simplify the interface.

The changes also update the help message to remove the reference to the "Join w/Ping" button, as it is no longer available.