feat: add modal for requesting Egg Inc ID
Adds a new function `RequestEggIncIDModal` that sends a modal to the user
requesting their Egg Inc ID. This modal is used when the user's Egg Inc ID
is not found or is invalid.
    
The `HandleEggIDModalSubmit` function is also added to handle the submission
of the modal, validating the Egg Inc ID and storing the encrypted ID in the
user's settings.