feat(boost/virtue): Add TruthEggBreakpoints table
Adds a new `init()` function that builds a table of breakpoints for Truth
Eggs above 16. This table is used to calculate the required eggs laid for
Truth Eggs above 16.

The changes also remove the hardcoded breakpoint values for Truth Eggs
above 100q, as these can now be calculated using the new table.